# demo-openapi.yaml
openapi: 3.0.0

info:
  title: Demo API
  description: API for demo project.
  version: 0.1.0
  
servers:
  - url: http://localhost:6868
paths:
  /users:
    get:
      summary: Returns a list of users.
      responses:
        '200':    # status code
          description: A JSON array of users
          content:
            application/json:
              schema: 
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Creates a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserNoId'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /users/{id}:
    get:
      summary: Gets user by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: string
    put:
      summary: Replaces user by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserNoId'
      responses:
        '200':
          description: Putted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      summary: Deletes user by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: Deleted successfully

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        age:
          type: integer
    UserNoId:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
    Id:
      type: integer
      format: int64